Enable deployment of multiple private chaincodes (#86) (#108)
Before private chaincodes had to be installed under the name `ecc`. This
limited the max number of deployed private chaincodes to one. With this
PR this limitation is fixed and multiple private chaincodes and "norma"
chaincodes can be installed and run at the same time. Note that private
chaincodes have to be installed with `ecc_` prefix but this currently
hidden through our peer wrapper. This prefix requirements may go away in
the future.

Signed-off-by: Marcus Brandenburger <bur@zurich.ibm.com>