Mock chaincode testing
This commit introduces a fpc chaincode mock server that exposes the
chaincode interface as a simple REST interface. Internally, this testing
server relies on shim.MockStub. The server can load normal go chaincodes
and fpc chaincodes by using ecc. This required a small refactoring of
ecc.

This mock server will be used to simplify fpc auction demo development,
in particular it allows rapid testing of the actual fpc chaincode and
the UI without the need to run an actual fabric network.

Moreover, this commit provides a simple mock implementation of the
auction written in go.

Signed-off-by: Marcus Brandenburger <bur@zurich.ibm.com>